+++ src/main/java/de/alpharogroup/crypto/core/AbstractCryptor.java
@@ -46,6 +46,7 @@ import de.alpharogroup.crypto.factories.AlgorithmParameterSpecFactory;
 import de.alpharogroup.crypto.factories.SecretKeyFactoryExtensions;
 import de.alpharogroup.crypto.model.CryptModel;
 import lombok.Getter;
+import lombok.NonNull;
 
 /**
  * The abstract class {@link AbstractCryptor} provides factory methods that may or must be
@@ -55,11 +56,13 @@ import lombok.Getter;
  *            the generic type of the cipher
  * @param <K>
  *            the generic type of the key
+ * @param <T>
+ *            the generic type of the decorator objects
  *
  * @author Asterios Raptis
  * @version 1.0
  */
-public abstract class AbstractCryptor<C, K> implements Serializable, Cryptor
+public abstract class AbstractCryptor<C, K, T> implements Serializable, Cryptor
 {
 
 	/** The Constant serialVersionUID. */
@@ -67,7 +70,7 @@ public abstract class AbstractCryptor<C, K> implements Serializable, Cryptor
 
 	/** The crypto model. */
 	@Getter
-	protected final CryptModel<C, K> model;
+	protected final CryptModel<C, K, T> model;
 
 	/**
 	 * Constructor with the given {@link CryptModel}.
@@ -89,11 +92,11 @@ public abstract class AbstractCryptor<C, K> implements Serializable, Cryptor
 	 * @throws UnsupportedEncodingException
 	 *             is thrown if the named charset is not supported.
 	 */
-	public AbstractCryptor(final CryptModel<C, K> model)
+	public AbstractCryptor(final @NonNull CryptModel<C, K, T> model)
 		throws InvalidKeyException, NoSuchAlgorithmException, InvalidKeySpecException,
 		NoSuchPaddingException, InvalidAlgorithmParameterException, UnsupportedEncodingException
 	{
-		Check.get().notNull(model, "model").notNull(model.getKey(), "model.getKey()");
+		Check.get().notNull(model.getKey(), "model.getKey()");
 		this.model = model;
 		onInitialize();
 	}
@@ -123,7 +126,7 @@ public abstract class AbstractCryptor<C, K> implements Serializable, Cryptor
 		NoSuchPaddingException, InvalidAlgorithmParameterException, UnsupportedEncodingException
 	{
 		Check.get().notNull(key, "key");
-		model = CryptModel.<C, K> builder().key(key).build();
+		model = CryptModel.<C, K, T> builder().key(key).build();
 		onInitialize();
 	}
 
@@ -135,11 +138,11 @@ public abstract class AbstractCryptor<C, K> implements Serializable, Cryptor
 	 */
 	protected String newAlgorithm()
 	{
-		if (getModel().getAlgorithm() == null)
+		if (model.getAlgorithm() == null)
 		{
 			return SunJCEAlgorithm.PBEWithMD5AndDES.getAlgorithm();
 		}
-		return getModel().getAlgorithm().getAlgorithm();
+		return model.getAlgorithm().getAlgorithm();
 	}
 
 	/**
@@ -230,11 +233,11 @@ public abstract class AbstractCryptor<C, K> implements Serializable, Cryptor
 	 */
 	protected int newIterationCount()
 	{
-		if (getModel().getIterationCount() == null)
+		if (model.getIterationCount() == null)
 		{
 			return CompoundAlgorithm.ITERATIONCOUNT;
 		}
-		return getModel().getIterationCount();
+		return model.getIterationCount();
 	}
 
 	/**