+++ src/main/java/de/alpharogroup/crypto/file/FileEncryptor.java
@@ -27,13 +27,13 @@ package de.alpharogroup.crypto.file;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
-import java.io.InputStream;
 import java.io.OutputStream;
 import java.io.UnsupportedEncodingException;
 import java.security.InvalidAlgorithmParameterException;
 import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
 import java.security.spec.InvalidKeySpecException;
+import java.util.List;
 
 import javax.crypto.Cipher;
 import javax.crypto.NoSuchPaddingException;
@@ -41,8 +41,10 @@ import javax.crypto.NoSuchPaddingException;
 import org.apache.commons.io.FilenameUtils;
 
 import de.alpharogroup.crypto.core.AbstractFileEncryptor;
+import de.alpharogroup.crypto.decorator.CryptObjectDecoratorExtensions;
 import de.alpharogroup.crypto.io.CryptoCipherInputStream;
 import de.alpharogroup.crypto.model.CryptModel;
+import de.alpharogroup.crypto.model.CryptObjectDecorator;
 
 /**
  * The class {@link FileEncryptor} can encrypt files with the given crypt model.
@@ -74,7 +76,7 @@ public class FileEncryptor extends AbstractFileEncryptor
 	 * @throws UnsupportedEncodingException
 	 *             is thrown if the named charset is not supported.
 	 */
-	public FileEncryptor(final CryptModel<Cipher, String> model)
+	public FileEncryptor(final CryptModel<Cipher, String, String> model)
 		throws InvalidKeyException, NoSuchAlgorithmException, InvalidKeySpecException,
 		NoSuchPaddingException, InvalidAlgorithmParameterException, UnsupportedEncodingException
 	{
@@ -103,7 +105,7 @@ public class FileEncryptor extends AbstractFileEncryptor
 	 * @throws UnsupportedEncodingException
 	 *             is thrown if the named charset is not supported.
 	 */
-	public FileEncryptor(final CryptModel<Cipher, String> model, final File encryptedFile)
+	public FileEncryptor(final CryptModel<Cipher, String, String> model, final File encryptedFile)
 		throws InvalidKeyException, NoSuchAlgorithmException, InvalidKeySpecException,
 		NoSuchPaddingException, InvalidAlgorithmParameterException, UnsupportedEncodingException
 	{
@@ -122,27 +124,30 @@ public class FileEncryptor extends AbstractFileEncryptor
 			final String filename = FilenameUtils.getBaseName(toEncrypt.getName());
 			encryptedFile = newEncryptedFile(toEncrypt.getParent(), filename + ".enc");
 		}
-
-		final InputStream fis = new FileInputStream(toEncrypt);
-		final CryptoCipherInputStream cis = new CryptoCipherInputStream(fis,
-			getModel().getCipher());
-
-		final OutputStream out = new FileOutputStream(encryptedFile);
-
-		int c;
-
-		while ((c = cis.read()) != -1)
+		List<CryptObjectDecorator<String>> decorators = getModel().getDecorators();
+		if (decorators != null && !decorators.isEmpty())
 		{
-			out.write(c);
+			for (int i = 0; i < decorators.size(); i++)
+			{
+				CryptObjectDecoratorExtensions.decorateFile(toEncrypt, decorators.get(i));
+			}
 		}
-
-		cis.close();
-		out.close();
+		try (
+			CryptoCipherInputStream cis = new CryptoCipherInputStream(
+				new FileInputStream(toEncrypt), getModel().getCipher());
+			OutputStream out = new FileOutputStream(encryptedFile))
+		{
+			int c;
+			while ((c = cis.read()) != -1)
+			{
+				out.write(c);
+			}
+		}
 		return encryptedFile;
 	}
 
 	/**
-	 * 
+	 *
 	 * Factory method for creating the new decrypted {@link File} if it is not exists. This method
 	 * is invoked in the constructor from the derived classes and can be overridden so users can
 	 * provide their own version of creating the new decrypted {@link File}