+++ src/test/java/de/alpharogroup/crypto/obfuscation/character/ObfuscatorExtensionsTest.java
@@ -27,22 +27,31 @@ package de.alpharogroup.crypto.obfuscation.character;
 import static org.testng.AssertJUnit.assertEquals;
 
 import java.io.File;
+import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
+import java.util.List;
 import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
 
+import org.apache.commons.codec.DecoderException;
 import org.meanbean.test.BeanTestException;
 import org.meanbean.test.BeanTester;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
 
 import com.google.common.collect.BiMap;
+import com.google.common.collect.HashBiMap;
 import com.thoughtworks.xstream.XStream;
 
 import de.alpharogroup.AbstractTestCase;
 import de.alpharogroup.collections.map.MapFactory;
 import de.alpharogroup.collections.pairs.KeyValuePair;
+import de.alpharogroup.collections.set.SetFactory;
 import de.alpharogroup.crypto.file.xml.XmlDecryptionExtensions;
+import de.alpharogroup.crypto.obfuscation.rule.CharacterObfuscationOperationRule;
 import de.alpharogroup.crypto.obfuscation.rule.ObfuscationOperationRule;
+import de.alpharogroup.crypto.obfuscation.rule.Operation;
 import de.alpharogroup.file.search.PathFinder;
 import lombok.AccessLevel;
 import lombok.experimental.FieldDefaults;
@@ -56,11 +65,17 @@ public class ObfuscatorExtensionsTest extends AbstractTestCase<String, String>
 
 	/** The aliases for the {@link XStream} object */
 	Map<String, Class<?>> aliases;
+	List<KeyValuePair<Character, ObfuscationOperationRule<Character, Character>>> data;
 	BiMap<Character, ObfuscationOperationRule<Character, Character>> fullSizeRules;
 	BiMap<Character, ObfuscationOperationRule<Character, Character>> smallSizeRules;
 	String stringToDisentangle;
-
 	String stringToObfuscate;
+
+	BiMap<Character, ObfuscationOperationRule<Character, Character>> testRules;
+	File xmlDir;
+	File xmlFile;
+
+	File xmlListFile;
 	/** The {@link XStream} object */
 	XStream xStream;
 	{
@@ -70,6 +85,16 @@ public class ObfuscatorExtensionsTest extends AbstractTestCase<String, String>
 		aliases = MapFactory.newLinkedHashMap();
 		aliases.put("KeyValuePair", KeyValuePair.class);
 		aliases.put("ObfuscationOperationRule", ObfuscationOperationRule.class);
+	}
+
+	BiMap<Character, ObfuscationOperationRule<Character, Character>> loadXmlListToBiMap(
+		File xmlListFile) throws IOException, DecoderException
+	{
+		List<KeyValuePair<Character, ObfuscationOperationRule<Character, Character>>> data = XmlDecryptionExtensions
+			.readFromFileAsXmlAndHex(xStream, aliases, xmlListFile);
+		BiMap<Character, ObfuscationOperationRule<Character, Character>> rules = HashBiMap
+			.create(KeyValuePair.toMap(data));
+		return rules;
 	}
 
 	/**
@@ -83,8 +108,6 @@ public class ObfuscatorExtensionsTest extends AbstractTestCase<String, String>
 	protected void setUp() throws Exception
 	{
 		super.setUp();
-		File xmlFile;
-		File xmlDir;
 
 		xmlDir = new File(PathFinder.getSrcTestResourcesDir(), "xml");
 		xmlFile = new File(xmlDir, "small-size-operation-rules.oor");
@@ -93,6 +116,10 @@ public class ObfuscatorExtensionsTest extends AbstractTestCase<String, String>
 
 		xmlFile = new File(xmlDir, "full-size-operation-rules.oor");
 		fullSizeRules = XmlDecryptionExtensions.readFromFileAsXmlAndHex(xStream, aliases, xmlFile);
+
+		xmlListFile = new File(xmlDir, "pwhex-foo.oor");
+		data = XmlDecryptionExtensions.readFromFileAsXmlAndHex(xStream, aliases, xmlListFile);
+		testRules = loadXmlListToBiMap(xmlListFile);
 
 	}
 
@@ -139,6 +166,97 @@ public class ObfuscatorExtensionsTest extends AbstractTestCase<String, String>
 	}
 
 	/**
+	 * Test method for {@link ObfuscatorExtensions#disentangle(BiMap, String)}
+	 *
+	 * @throws DecoderException
+	 * @throws IOException
+	 */
+	@Test
+	public void testDisentangleImprovedWithTestRules() throws IOException, DecoderException
+	{
+
+		String obfuscateWith;
+		// new scenario...
+		stringToDisentangle = "Ast3r70s";
+
+		actual = ObfuscatorExtensions.disentangleImproved(testRules, stringToDisentangle);
+		expected = "asterios";
+		assertEquals(expected, actual);
+
+		// new scenario...
+		stringToObfuscate = "Asterios";
+
+		testRules = loadXmlListToBiMap(xmlListFile);
+		testRules.put(Character.valueOf('A'), ObfuscationOperationRule
+			.<Character, Character> builder().character('A').replaceWith('5').build());
+		obfuscateWith = ObfuscatorExtensions.obfuscateWith(testRules, stringToObfuscate);
+
+		actual = ObfuscatorExtensions.disentangleImproved(testRules, obfuscateWith);
+		expected = stringToObfuscate;
+		assertEquals(expected, actual);
+
+	}
+
+	/**
+	 * Test method for {@link ObfuscatorExtensions#disentangle(BiMap, String)}
+	 */
+	@Test
+	public void testDisentangleWithTestRules()
+	{
+		String obfuscateWith;
+		// new scenario...
+		stringToObfuscate = "asterios";
+		obfuscateWith = ObfuscatorExtensions.obfuscateWith(testRules, stringToObfuscate);
+
+		actual = ObfuscatorExtensions.disentangle(data, obfuscateWith);
+		expected = stringToObfuscate;
+		assertEquals(expected, actual);
+
+		// new scenario...
+		stringToObfuscate = "Leonardo!";
+		obfuscateWith = ObfuscatorExtensions.obfuscateWith(testRules, stringToObfuscate);
+
+		actual = ObfuscatorExtensions.disentangle(data, obfuscateWith);
+		expected = stringToObfuscate;
+		assertEquals(expected, actual);
+
+	}
+
+	@Test
+	public void testInverse()
+	{
+		Character character;
+		Set<Integer> indexes;
+		boolean inverted;
+		Operation operation;
+		Character replaceWith;
+		CharacterObfuscationOperationRule rule;
+		Character actual;
+		Character expected;
+		Optional<Character> operatedCharacter;
+
+		inverted = false;
+		operatedCharacter = Optional.empty();
+		character = Character.valueOf('a');
+		replaceWith = 'b';
+		operation = Operation.UPPERCASE;
+		indexes = SetFactory.newHashSet(0, 2);
+
+		rule = new CharacterObfuscationOperationRule(character, indexes, inverted,
+			operatedCharacter, operation, replaceWith);
+
+		ObfuscatorExtensions.inverse(rule);
+
+		actual = rule.getCharacter();
+		expected = replaceWith;
+		assertEquals(actual, expected);
+
+		actual = rule.getReplaceWith();
+		expected = character;
+		assertEquals(actual, expected);
+	}
+
+	/**
 	 * Test method for {@link ObfuscatorExtensions#obfuscateWith(BiMap, String)}
 	 */
 	@Test
@@ -155,6 +273,12 @@ public class ObfuscatorExtensionsTest extends AbstractTestCase<String, String>
 
 		actual = ObfuscatorExtensions.obfuscateWith(smallSizeRules, stringToObfuscate);
 		expected = "AcACd";
+		assertEquals(expected, actual);
+
+		// new scenario...
+		stringToObfuscate = "asterios";
+		actual = ObfuscatorExtensions.obfuscateWith(testRules, stringToObfuscate);
+		expected = "Ast3r70s";
 		assertEquals(expected, actual);
 	}